==> 12_PyCode/Stage02_AN_PdpContributionEstimation_v0_Execute.ipynb <==
# Hyperparameter Search

%pwd

%cd ..

## Import Package 

from glob import glob
from joblib import dump, load
import joblib
import numpy as np
import pandas as pd
import random
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error, r2_score
from sklearn.model_selection import train_test_split, KFold

### Function

### X and y
def getXandY(Output_Vari):
    y_list = glob("/home/pj24001881/ku40001335/Greenness_NighttimeLight_WB/01_Data/*_y_" + Output_Vari + "*.csv")
    y = pd.read_csv(y_list[0], index_col=0)
    y = y.iloc[:,0].to_numpy()
    X_list = glob("/home/pj24001881/ku40001335/Greenness_NighttimeLight_WB/01_Data/*_X_" + Output_Vari + "*.csv")
    X = pd.read_csv(X_list[0], index_col=0)
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, 
                                                        random_state=1)
    return X_train, X_test, y_train, y_test

def XYSplit(model):
    total_tree_number = model.n_estimators
    X_split_array = []
    Y_split_array = []
    for tree_order in list(range(total_tree_number)):
        feature = model.estimators_[tree_order].tree_.feature
        threshold = model.estimators_[tree_order].tree_.threshold
        
        df = pd.concat([pd.Series(feature), pd.Series(threshold)], axis=1)
        df.columns = ['feature', 'threshold']
        X_split = df[(df['feature']==2)|(df['feature']==-27)]
        Y_split = df[(df['feature']==1)|(df['feature']==-28)]
        
        X_split = X_split.sort_values('threshold')
        Y_split = Y_split.sort_values('threshold')
        X_split = np.array(X_split.threshold, dtype='float64')
        Y_split = np.array(Y_split.threshold, dtype='float64')
        
        X_split_array.append(X_split)
        Y_split_array.append(Y_split)
        
    return (X_split_array, Y_split_array)

def findBoundaryArray(split_array, data_degree):
    # Use joblib to parallelize the computation
    boundary_array = joblib.Parallel(n_jobs=60)(
        joblib.delayed(findBoundary)(split_array, observation)
        for observation in data_degree)
    
    return boundary_array

def findBoundary(split_array, observation):
    before_array = []
    after_array = []
    for split in split_array:
        split_add = np.insert(split, 0, observation)
        split_add = np.sort(split_add)
        location = np.where(split_add == observation)[0][0]
        if location == 0:
            before = observation - 1
        else:
            before = split[location - 1]
        if location == len(split_add)-1:
            after = observation + 1
        else:
            after = split[location]
        before_array.append(before)
        after_array.append(after)
    #before_observation = np.min(np.array(before_array))
    #after_observation = np.max(np.array(after_array))
    before_observation = np.median(np.array(before_array))
    after_observation = np.median(np.array(after_array))
    #before_observation = np.quantile(np.array(before_array), 0.75)
    #after_observation = np.quantile(np.array(after_array), 0.25)
    return [before_observation, after_observation]

import multiprocessing
def findBoundary(split_array, observation):
    before_array = []
    after_array = []
    
    for split in split_array:
        idx = np.searchsorted(split, observation)
        before = split[idx - 1] if idx > 0 else observation - 1
        after = split[idx] if idx < len(split) else observation + 1
        before_array.append(before)
        after_array.append(after)

    before_observation = np.median(np.array(before_array))
    after_observation = np.median(np.array(after_array))
    #before_observation = np.quantile(before_array, 0.75)
    #after_observation = np.quantile(after_array, 0.25)
    
    return [before_observation, after_observation]

def worker(args):
    return findBoundary(*args)

def findBoundaryArray(split_array, data_degree):
    # Use a pool of workers to execute processes in parallel
    with multiprocessing.Pool(processes=multiprocessing.cpu_count()) as pool:
        # Create a list of arguments for each process
        args = [(split_array, observation) for observation in data_degree]
        # Map the worker function over the list of arguments
        boundary_array = pool.map(worker, args)
    return boundary_array

## Experiment for Happiness

### Load Data

Output_Vari = "Happinessoverall"

X_train, X_test, y_train, y_test = getXandY(Output_Vari)

X_train.head()

X_train.describe()

X_train.columns

X_train.shape

X_test.shape

X = pd.concat([X_train, X_test])

X.shape

y = np.concatenate([y_train, y_test])

y.shape

X.describe()

X.columns

### Run Test Model

model = RandomForestRegressor(n_estimators = 3900, min_samples_split = 8, max_features = 0.4, max_samples = 0.5, max_depth = 16, n_jobs = -1)
model.fit(X_train, y_train)

y_pred = model.predict(X_test)
r2_score(y_test, y_pred)

X_test.shape

X_pdp_list = []
for incre in list(range(0, 11)):
    X_test_change = X_test.copy()
    X_test_change['NDVI'] = X_test_change['NDVI'] + 0.1 * incre
    X_pdp_list.append(X_test_change)
X_pdp = pd.concat(X_pdp_list)

X_pdp.shape

y_pdp = model.predict(X_pdp)

split_list = XYSplit(model)

len(split_list[0])

X_split_array, Y_split_array = XYSplit(model)

%%time
leftRightBoundary = findBoundaryArray(X_split_array, X_test.iloc[:10_000,2])
upDownBoundary = findBoundaryArray(Y_split_array, X_test.iloc[:10_000,1])

len(leftRightBoundary)

X_test_1000 = X_test.iloc[:10_000,:]

X_test_1000['left'] = np.array(leftRightBoundary)[:,0]
X_test_1000['right'] = np.array(leftRightBoundary)[:,1]
X_test_1000['down'] = np.array(upDownBoundary)[:,0]
X_test_1000['up'] = np.array(upDownBoundary)[:,1]

X_test_1000

## Estimated PDP

kf = KFold(n_splits=10, shuffle=True, random_state=42)

### NDVI

run = False
if run:
    pdp_df_list = []
    
    i = 1
    for train_index, test_index in kf.split(X):
        print(f"This is {i} fold")
        X_train, X_test = X.iloc[train_index, :], X.iloc[test_index, :]
        y_train, y_test = y[train_index], y[test_index]
        
        # Train the model
        model = RandomForestRegressor(n_estimators = 3900, min_samples_split = 8, max_features = 0.4, max_samples = 0.5, max_depth = 16, n_jobs = -1)
        model.fit(X_train, y_train)
    
        X_pdp_list = []
        for incre in list(range(0, 41)):
            X_test_change = X_test.copy()
            X_test_change['NDVI'] = X_test_change['NDVI'] + 0.05 * incre
    
            X_split_array, Y_split_array = XYSplit(model)
            leftRightBoundary = findBoundaryArray(X_split_array, X_test.iloc[:,2])
            upDownBoundary = findBoundaryArray(Y_split_array, X_test.iloc[:,1])
            
            X_test_change['left'] = np.array(leftRightBoundary)[:,0]
            X_test_change['right'] = np.array(leftRightBoundary)[:,1]
            X_test_change['down'] = np.array(upDownBoundary)[:,0]
            X_test_change['up'] = np.array(upDownBoundary)[:,1]
            
            X_pdp_list.append(X_test_change)
        X_pdp = pd.concat(X_pdp_list)
        y_pdp = model.predict(X_pdp.iloc[:,:-4])
    
        X_pdp['Pred_y'] = y_pdp
    
        X_pdp.to_parquet(f'/home/pj24001881/ku40001335/Greenness_NighttimeLight_WB/03_Results/PdpLocal/PDP_Local_NDVI_{i}.parquet')
    
        pdp_df_list.append(X_pdp)
        i = i + 1
    
    pdp_df = pd.concat(pdp_df_list)

### NTL

run = False
if run:
    pdp_df_list = []
    
    i = 1
    for train_index, test_index in kf.split(X):
        print(f"This is {i} fold")
        X_train, X_test = X.iloc[train_index, :], X.iloc[test_index, :]
        y_train, y_test = y[train_index], y[test_index]
        
        # Train the model
        model = RandomForestRegressor(n_estimators = 3900, min_samples_split = 8, max_features = 0.4, max_samples = 0.5, max_depth = 16, n_jobs = -1)
        model.fit(X_train, y_train)
    
        X_pdp_list = []
        for incre in list(range(0, 41)):
            X_test_change = X_test.copy()
            X_test_change['NTL'] = X_test_change['NTL'] + 0.05 * incre
    
            X_split_array, Y_split_array = XYSplit(model)
            leftRightBoundary = findBoundaryArray(X_split_array, X_test.iloc[:,2])
            upDownBoundary = findBoundaryArray(Y_split_array, X_test.iloc[:,1])
            
            X_test_change['left'] = np.array(leftRightBoundary)[:,0]
            X_test_change['right'] = np.array(leftRightBoundary)[:,1]
            X_test_change['down'] = np.array(upDownBoundary)[:,0]
            X_test_change['up'] = np.array(upDownBoundary)[:,1]
            
            X_pdp_list.append(X_test_change)
        X_pdp = pd.concat(X_pdp_list)
        y_pdp = model.predict(X_pdp.iloc[:,:-4])
    
        X_pdp['Pred_y'] = y_pdp
    
        X_pdp.to_parquet(f'/home/pj24001881/ku40001335/Greenness_NighttimeLight_WB/03_Results/PdpLocal/PDP_Local_NTL_{i}.parquet')
    
        pdp_df_list.append(X_pdp)
        i = i + 1
    
    pdp_df = pd.concat(pdp_df_list)

### Income

pdp_df_list = []

i = 1
for train_index, test_index in kf.split(X):
    print(f"This is {i} fold")
    X_train, X_test = X.iloc[train_index, :], X.iloc[test_index, :]
    y_train, y_test = y[train_index], y[test_index]
    
    # Train the model
    model = RandomForestRegressor(n_estimators = 3900, min_samples_split = 8, max_features = 0.4, max_samples = 0.5, max_depth = 16, n_jobs = -1)
    model.fit(X_train, y_train)

    X_pdp_list = []
    for incre in list(range(0, 41)):
        X_test_change = X_test.copy()
        X_test_change['income_indiv'] = X_test_change['income_indiv'] + 0.05 * incre

        X_split_array, Y_split_array = XYSplit(model)
        leftRightBoundary = findBoundaryArray(X_split_array, X_test.iloc[:,2])
        upDownBoundary = findBoundaryArray(Y_split_array, X_test.iloc[:,1])
        
        X_test_change['left'] = np.array(leftRightBoundary)[:,0]
        X_test_change['right'] = np.array(leftRightBoundary)[:,1]
        X_test_change['down'] = np.array(upDownBoundary)[:,0]
        X_test_change['up'] = np.array(upDownBoundary)[:,1]
        
        X_pdp_list.append(X_test_change)
    X_pdp = pd.concat(X_pdp_list)
    y_pdp = model.predict(X_pdp.iloc[:,:-4])

    X_pdp['Pred_y'] = y_pdp

    X_pdp.to_parquet(f'/home/pj24001881/ku40001335/Greenness_NighttimeLight_WB/03_Results/PdpLocal/PDP_Local_income_indiv_{i}.parquet')

    pdp_df_list.append(X_pdp)
    i = i + 1

pdp_df = pd.concat(pdp_df_list)
